requirements.txt
google-cloud-datastore==1.12.0
Flask==1.1.2
flask-marshmallow==0.12.0
